Web Demo/requirements.txt
attrs==19.1.0
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
numpy==1.16.3
opencv-python==4.1.0.25
pandas==0.24.2
Pillow==6.0.0
pytesseract==0.2.6
python-dateutil==2.8.0
pytz==2019.1
scipy==1.2.1
six==1.12.0
werkzeug==0.14.1
jsonpickle==1.
tornado==6.0.2
Shapely==1.6.4.post2
bs4
